requirements.txt
ultralytics
streamlit
roboflow
pillow
opencv-python
numpy